% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tabela_fotos}
\alias{tabela_fotos}
\title{Discursos proferidos na CPI da Pandemia no Senado.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 113 rows and 2 columns.
}
\usage{
tabela_fotos
}
\description{
Base raspada e organizada do site do Senado Federal com todos os discursos
proferidos nas reuniões da CPI da Pandemia. Os discursos foram transcritos
pela equipe do Senado Federal e disponibilizados no site.
}
\keyword{datasets}
